+++ src/SDLApplication/SDLStdioLogger.h
@@ -14,7 +14,7 @@ namespace Mengine
         ~SDLStdioLogger() override;
 
     public:
-        void log( ELoggerLevel _level, uint32_t _flag, const Char * _data, uint32_t _count ) override;
+        void log( ELoggerLevel _level, uint32_t _flag, const Char * _data, size_t _count ) override;
         void flush() override;
     };
 }
